# Skip barriers of noops